fix: lazy load code splitting
