+++ src/preview/index.ts
@@ -1,9 +1,13 @@
+import {lazy} from 'react'
+
 import {PreviewMode} from './PreviewMode'
 
 export default PreviewMode
 export * from './PreviewMode'
 export * from './PreviewSubscription'
-export * from './PreviewSubscriptionWithToken'
+export type {PreviewSubscriptionWithTokenProps} from './PreviewSubscriptionWithToken'
 export * from './useAuthenticated'
 export * from './useGroqStore'
 export * from './useSyncGroqStore'
+
+export const PreviewSubscriptionWithToken = lazy(() => import('./PreviewSubscriptionWithToken'))
